# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Homeschool Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CourseWithCounts {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  gradeLevel?: string | null;
  published: boolean;
  ownerId: string;
  createdAt: string;
  updatedAt: string;
  unitCount: number;
  lessonCount: number;
}

export interface DashboardSummary {
  courseCount: number;
  lessonCount: number;
  recentCourses: CourseWithCounts[];
}

export interface Course {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  gradeLevel?: string | null;
  published: boolean;
  ownerId: string;
  createdAt: string;
  updatedAt: string;
}

export interface Lesson {
  id: string;
  title: string;
  content: string;
  order: number;
  published: boolean;
  unitId: string;
  createdAt: string;
  updatedAt: string;
}

export interface UnitWithLessons {
  id: string;
  title: string;
  order: number;
  courseId: string;
  createdAt: string;
  updatedAt: string;
  lessons: Lesson[];
}

export interface CourseDetail {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  gradeLevel?: string | null;
  published: boolean;
  ownerId: string;
  createdAt: string;
  updatedAt: string;
  units: UnitWithLessons[];
}

export interface CourseInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  subject?: string;
  gradeLevel?: string;
}

export interface CourseUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  gradeLevel?: string | null;
}

export interface PublishToggle {
  published: boolean;
}

export interface Unit {
  id: string;
  title: string;
  order: number;
  courseId: string;
  createdAt: string;
  updatedAt: string;
}

export interface UnitInput {
  /** @minLength 1 */
  title: string;
}

export interface LessonDetail {
  id: string;
  title: string;
  content: string;
  order: number;
  published: boolean;
  unitId: string;
  courseId: string;
  courseTitle: string;
  unitTitle: string;
  createdAt: string;
  updatedAt: string;
}

export interface LessonInput {
  /** @minLength 1 */
  title: string;
}

export interface LessonUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  published?: boolean;
}