# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Homeschool Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: courses
    description: Course operations
  - name: units
    description: Unit operations
  - name: lessons
    description: Lesson operations
  - name: dashboard
    description: Dashboard summary operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseWithCounts"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by ID with units and lessons
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /courses/{id}/publish:
    patch:
      operationId: toggleCoursePublished
      tags: [courses]
      summary: Toggle published status of a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishToggle"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

  /courses/{id}/units:
    post:
      operationId: createUnit
      tags: [units]
      summary: Add a unit to a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"

  /units/{unitId}:
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Delete a unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /units/{unitId}/lessons:
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Add a lesson to a unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a lesson by ID
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lesson detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update a lesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonUpdate"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete a lesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    DashboardSummary:
      type: object
      properties:
        courseCount:
          type: number
        lessonCount:
          type: number
        recentCourses:
          type: array
          items:
            $ref: "#/components/schemas/CourseWithCounts"
      required:
        - courseCount
        - lessonCount
        - recentCourses

    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        gradeLevel:
          type: ["string", "null"]
        published:
          type: boolean
        ownerId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - published
        - ownerId
        - createdAt
        - updatedAt

    CourseWithCounts:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        gradeLevel:
          type: ["string", "null"]
        published:
          type: boolean
        ownerId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        unitCount:
          type: number
        lessonCount:
          type: number
      required:
        - id
        - title
        - published
        - ownerId
        - createdAt
        - updatedAt
        - unitCount
        - lessonCount

    CourseDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        gradeLevel:
          type: ["string", "null"]
        published:
          type: boolean
        ownerId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        units:
          type: array
          items:
            $ref: "#/components/schemas/UnitWithLessons"
      required:
        - id
        - title
        - published
        - ownerId
        - createdAt
        - updatedAt
        - units

    CourseInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        subject:
          type: string
        gradeLevel:
          type: string
      required:
        - title

    CourseUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        gradeLevel:
          type: ["string", "null"]

    PublishToggle:
      type: object
      properties:
        published:
          type: boolean
      required:
        - published

    Unit:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        order:
          type: number
        courseId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - order
        - courseId
        - createdAt
        - updatedAt

    UnitWithLessons:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        order:
          type: number
        courseId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
      required:
        - id
        - title
        - order
        - courseId
        - createdAt
        - updatedAt
        - lessons

    UnitInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
      required:
        - title

    Lesson:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        order:
          type: number
        published:
          type: boolean
        unitId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - content
        - order
        - published
        - unitId
        - createdAt
        - updatedAt

    LessonDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        order:
          type: number
        published:
          type: boolean
        unitId:
          type: string
        courseId:
          type: string
        courseTitle:
          type: string
        unitTitle:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - content
        - order
        - published
        - unitId
        - courseId
        - courseTitle
        - unitTitle
        - createdAt
        - updatedAt

    LessonInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
      required:
        - title

    LessonUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        published:
          type: boolean