# lib/api-zod/src/generated/types/courseDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Homeschool Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UnitWithLessons } from './unitWithLessons';

export interface CourseDetail {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  gradeLevel?: string | null;
  published: boolean;
  ownerId: string;
  createdAt: string;
  updatedAt: string;
  units: UnitWithLessons[];
}