# lib/api-zod/src/generated/types/courseUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Homeschool Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CourseUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  gradeLevel?: string | null;
}