# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Homeschool Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CourseWithCounts } from './courseWithCounts';

export interface DashboardSummary {
  courseCount: number;
  lessonCount: number;
  recentCourses: CourseWithCounts[];
}