# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Homeschool Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: string;
  title: string;
  content: string;
  order: number;
  published: boolean;
  unitId: string;
  createdAt: string;
  updatedAt: string;
}