# lib/api-zod/src/generated/types/unitInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Homeschool Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UnitInput {
  /** @minLength 1 */
  title: string;
}