# lib/api-zod/src/generated/types/unitWithLessons.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Homeschool Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Lesson } from './lesson';

export interface UnitWithLessons {
  id: string;
  title: string;
  order: number;
  courseId: string;
  createdAt: string;
  updatedAt: string;
  lessons: Lesson[];
}